+++ app/Http/Controllers/SupplyController.php
@@ -44,6 +44,7 @@ use Adshares\Adserver\ViewModel\MediumName;
 use Adshares\Adserver\ViewModel\ZoneSize;
 use Adshares\Common\Application\Service\AdUser;
 use Adshares\Common\Domain\ValueObject\SecureUrl;
+use Adshares\Common\Domain\ValueObject\Uuid;
 use Adshares\Common\Domain\ValueObject\WalletAddress;
 use Adshares\Common\Exception\InvalidArgumentException;
 use Adshares\Common\Exception\RuntimeException;
@@ -477,7 +478,7 @@ class SupplyController extends Controller
 
         if ($foundBanners->exists(fn($key, $element) => $element != null)) {
             NetworkImpression::register(
-                Utils::hexUuidFromBase64UrlWithChecksum($impressionId),
+                self::impressionIdToUuid($impressionId),
                 Utils::hexUuidFromBase64UrlWithChecksum($tid),
                 $impressionContext,
                 $userContext,
@@ -579,7 +580,7 @@ class SupplyController extends Controller
     {
         $impressionId = $request->query->get('iid');
         $networkImpression = NetworkImpression::fetchByImpressionId(
-            Utils::hexUuidFromBase64UrlWithChecksum($impressionId)
+            self::impressionIdToUuid($impressionId)
         );
         if (null === $networkImpression || !$networkImpression->context->banner_id) {
             throw new NotFoundHttpException();
@@ -619,16 +620,20 @@ class SupplyController extends Controller
 
         $url = $this->addQueryStringToUrl($request, $url);
 
-        $caseId = $request->query->get('cid');
+        $caseId = str_replace('-', '', $request->query->get('cid'));
         if (null === ($networkCase = NetworkCase::fetchByCaseId($caseId))) {
             throw new NotFoundHttpException();
         }
 
         $payTo = AdsUtils::normalizeAddress(config('app.adshares_address'));
-        try {
-            $zoneId = ($networkCase->zone_id ?? null) ?: Utils::getZoneIdFromContext($request->query->get('ctx'));
-        } catch (RuntimeException $exception) {
-            throw new UnprocessableEntityHttpException($exception->getMessage(), $exception);
+        if (null === ($zoneId = ($networkCase->zone_id ?? null))) {
+            try {
+                $zoneId = null !== $request->query->get('zid')
+                    ? str_replace('-', '', $request->query->get('zid'))
+                    : Utils::getZoneIdFromContext($request->query->get('ctx'));
+            } catch (RuntimeException $exception) {
+                throw new UnprocessableEntityHttpException($exception->getMessage(), $exception);
+            }
         }
         $publisherId = Zone::fetchPublisherPublicIdByPublicId($zoneId);
         $impressionId = $request->query->get('iid');
@@ -636,6 +641,10 @@ class SupplyController extends Controller
         $url = Utils::addUrlParameter($url, 'pto', $payTo);
         $url = Utils::addUrlParameter($url, 'pid', $publisherId);
         $url = Utils::addUrlParameter($url, 'iid', $impressionId);
+        if (!$request->query->has('ctx')) {
+            $ctx = $this->buildCtx($networkCase->networkImpression, $impressionId, $zoneId);
+            $url = Utils::addUrlParameter($url, 'ctx', $ctx);
+        }
 
         $response = new RedirectResponse($url);
         $response->send();
@@ -693,7 +702,7 @@ class SupplyController extends Controller
     {
         $impressionId = $request->query->get('iid');
         $networkImpression = NetworkImpression::fetchByImpressionId(
-            Utils::hexUuidFromBase64UrlWithChecksum($impressionId)
+            self::impressionIdToUuid($impressionId)
         );
         if (null === $networkImpression || !$networkImpression->context->banner_id) {
             throw new NotFoundHttpException();
@@ -723,8 +732,11 @@ class SupplyController extends Controller
         $this->validateEventRequest($request);
 
         $impressionId = $request->query->get('iid');
+        if (null === $impressionId) {
+            throw new BadRequestHttpException('Invalid parameters.');
+        }
         $networkImpression = NetworkImpression::fetchByImpressionId(
-            Utils::hexUuidFromBase64UrlWithChecksum($impressionId)
+            self::impressionIdToUuid($impressionId)
         );
         if (null === $networkImpression) {
             throw new NotFoundHttpException();
@@ -735,22 +747,26 @@ class SupplyController extends Controller
             $url = $this->addQueryStringToUrl($request, $url);
         }
 
-        $caseId = $request->query->get('cid');
         $payTo = AdsUtils::normalizeAddress(config('app.adshares_address'));
 
-        try {
-            $zoneId = ($networkImpression->context->zone_id ?? null)
-                ?:
-                Utils::getZoneIdFromContext($request->query->get('ctx'));
-        } catch (RuntimeException $exception) {
-            throw new UnprocessableEntityHttpException($exception->getMessage(), $exception);
+        if (null === ($zoneId = ($networkImpression->context->zone_id ?? null))) {
+            try {
+                $zoneId = null !== $request->query->get('zid')
+                    ? str_replace('-', '', $request->query->get('zid'))
+                    : Utils::getZoneIdFromContext($request->query->get('ctx'));
+            } catch (RuntimeException $exception) {
+                throw new UnprocessableEntityHttpException($exception->getMessage(), $exception);
+            }
         }
         $publisherId = Zone::fetchPublisherPublicIdByPublicId($zoneId);
         $siteId = Zone::fetchSitePublicIdByPublicId($zoneId);
 
         $url = Utils::addUrlParameter($url, 'pto', $payTo);
         $url = Utils::addUrlParameter($url, 'pid', $publisherId);
-
+        if (!$request->query->has('ctx')) {
+            $ctx = $this->buildCtx($networkImpression, $impressionId, $zoneId);
+            $url = Utils::addUrlParameter($url, 'ctx', $ctx);
+        }
         $response = new RedirectResponse($url);
 
         if ($request->headers->has('Origin')) {
@@ -759,6 +775,7 @@ class SupplyController extends Controller
 
         $response->send();
 
+        $caseId = str_replace('-', '', $request->query->get('cid'));
         $networkCase = NetworkCase::create(
             $caseId,
             $publisherId,
@@ -778,8 +795,8 @@ class SupplyController extends Controller
     {
         if (
             !$request->query->has('r')
-            || !$request->query->has('ctx')
-            || !Utils::isUuidValid($request->query->get('cid'))
+            || !($request->query->has('ctx') || (Uuid::isValid($request->query->get('zid', ''))))
+            || !Uuid::isValid($request->query->get('cid', ''))
         ) {
             throw new BadRequestHttpException('Invalid parameters.');
         }
@@ -1227,5 +1244,37 @@ class SupplyController extends Controller
                 sprintf('Field `placements[].%s` must be a string', $field)
             );
         }
+    }
+
+    private static function impressionIdToUuid(string $impressionId): string
+    {
+        if (Uuid::isValid($impressionId)) {
+            return str_replace('-', '', $impressionId);
+        }
+        return Utils::hexUuidFromBase64UrlWithChecksum($impressionId);
+    }
+
+    private function buildCtx(
+        NetworkImpression $networkImpression,
+        string $impressionId,
+        string $zoneId,
+    ): string {
+        $impressionContext = $networkImpression->context;
+        $ctx = [
+            'page' => [
+                'iid' => $impressionId,
+                'keywords' => join(',', $impressionContext->site->keywords ?? []),
+                'metamask' => $impressionContext->device->extensions->metamask ?? 0,
+                'options' => '',
+                'pop' => $impressionContext->site->popup ?? 0,
+                'ref' => $impressionContext->site->referrer ?? '',
+                'url' => $impressionContext->site->page ?? '',
+                'zone' => $zoneId,
+            ]
+        ];
+        if (null !== ($inframe = $impressionContext->site->inframe)) {
+            $ctx['page']['frame'] = 'yes' === $inframe ? 1 : 0;
+        }
+        return Utils::encodeZones($ctx);
     }
 }