+++ tests/app/Http/Controllers/SupplyControllerTest.php
@@ -24,12 +24,15 @@ declare(strict_types=1);
 namespace Adshares\Adserver\Tests\Http\Controllers;
 
 use Adshares\Adserver\Client\GuzzleAdSelectClient;
+use Adshares\Adserver\Http\Utils;
 use Adshares\Adserver\Models\Banner;
 use Adshares\Adserver\Models\Config;
 use Adshares\Adserver\Models\NetworkBanner;
 use Adshares\Adserver\Models\NetworkCampaign;
 use Adshares\Adserver\Models\NetworkCase;
+use Adshares\Adserver\Models\NetworkCaseClick;
 use Adshares\Adserver\Models\NetworkHost;
+use Adshares\Adserver\Models\NetworkImpression;
 use Adshares\Adserver\Models\NetworkVectorsMeta;
 use Adshares\Adserver\Models\Site;
 use Adshares\Adserver\Models\User;
@@ -48,6 +51,7 @@ use Illuminate\Support\Facades\Crypt;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 use Psr\Http\Message\ResponseInterface;
+use Ramsey\Uuid\Uuid;
 use Symfony\Component\HttpFoundation\Response;
 
 final class SupplyControllerTest extends TestCase
@@ -611,6 +615,175 @@ final class SupplyControllerTest extends TestCase
         $response->assertJsonStructure(self::TARGETING_REACH_STRUCTURE);
     }
 
+    public function testLogNetworkClick(): void
+    {
+        [$query, $banner, $zone] = self::initBeforeLoggingClick();
+
+        $response = $this->get(self::buildLogClickUri($banner->uuid, $query));
+
+        $response->assertStatus(Response::HTTP_FOUND);
+        $response->assertHeader('Location');
+        $location = $response->headers->get('Location');
+        self::assertStringStartsWith('https://example.com/view', $location);
+        parse_str(parse_url($location, PHP_URL_QUERY), $locationQuery);
+        foreach (['cid', 'ctx', 'iid', 'pto', 'pid'] as $key) {
+            self::assertArrayHasKey($key, $locationQuery);
+        }
+        self::assertEquals('13245679801324567980132456798012', $locationQuery['cid']);
+        self::assertEquals('0001-00000005-CBCA', $locationQuery['pto']);
+        self::assertDatabaseCount(NetworkCaseClick::class, 1);
+    }
+
+    public function testLogNetworkClickWithoutContext(): void
+    {
+        [$query, $banner, $zone] = self::initBeforeLoggingClick();
+        unset($query['ctx']);
+        $query['zid'] = $zone->uuid;
+
+        $response = $this->get(self::buildLogClickUri($banner->uuid, $query));
+
+        $response->assertStatus(Response::HTTP_FOUND);
+        $response->assertHeader('Location');
+        $location = $response->headers->get('Location');
+        self::assertStringStartsWith('https://example.com/view', $location);
+        parse_str(parse_url($location, PHP_URL_QUERY), $locationQuery);
+        foreach (['cid', 'ctx', 'iid', 'pto', 'pid'] as $key) {
+            self::assertArrayHasKey($key, $locationQuery);
+        }
+        self::assertEquals('13245679801324567980132456798012', $locationQuery['cid']);
+        self::assertEquals('0001-00000005-CBCA', $locationQuery['pto']);
+        self::assertDatabaseCount(NetworkCaseClick::class, 1);
+    }
+
+    public function testLogNetworkClickFailWhileNoView(): void
+    {
+        [$query, $banner, $zone] = self::initBeforeLoggingView();
+
+        $response = $this->get(self::buildLogClickUri($banner->uuid, $query));
+
+        $response->assertStatus(Response::HTTP_NOT_FOUND);
+    }
+
+    public function testLogNetworkClickFailWhileInvalidRedirectUrlAndBannerId(): void
+    {
+        [$query, $banner, $zone] = self::initBeforeLoggingClick();
+        $query['r'] = '';
+
+        $response = $this->get(self::buildLogClickUri('invalid', $query));
+
+        $response->assertStatus(Response::HTTP_NOT_FOUND);
+    }
+
+    public function testLogNetworkView(): void
+    {
+        [$query, $banner, $zone] = self::initBeforeLoggingView();
+
+        $response = $this->get(self::buildLogViewUri($banner->uuid, $query));
+
+        $response->assertStatus(Response::HTTP_FOUND);
+        $response->assertHeader('Location');
+        $location = $response->headers->get('Location');
+        self::assertStringStartsWith('https://example.com/view', $location);
+        parse_str(parse_url($location, PHP_URL_QUERY), $locationQuery);
+        foreach (['cid', 'ctx', 'iid', 'pto', 'pid'] as $key) {
+            self::assertArrayHasKey($key, $locationQuery);
+        }
+        self::assertEquals('13245679801324567980132456798012', $locationQuery['cid']);
+        self::assertEquals('0001-00000005-CBCA', $locationQuery['pto']);
+    }
+
+    public function testLogNetworkViewWithoutContext(): void
+    {
+        [$query, $banner, $zone] = self::initBeforeLoggingView();
+        unset($query['ctx']);
+        $query['zid'] = $zone->uuid;
+
+        $response = $this->get(self::buildLogViewUri($banner->uuid, $query));
+
+        $response->assertStatus(Response::HTTP_FOUND);
+        $response->assertHeader('Location');
+        $location = $response->headers->get('Location');
+        self::assertStringStartsWith('https://example.com/view', $location);
+        parse_str(parse_url($location, PHP_URL_QUERY), $locationQuery);
+        foreach (['cid', 'ctx', 'iid', 'pto', 'pid'] as $key) {
+            self::assertArrayHasKey($key, $locationQuery);
+        }
+        self::assertEquals('13245679801324567980132456798012', $locationQuery['cid']);
+        self::assertEquals('0001-00000005-CBCA', $locationQuery['pto']);
+    }
+
+    public function testLogNetworkViewWhileCaseIdAndImpressionIdAreUuidV4(): void
+    {
+        [$query, $banner, $zone] = self::initBeforeLoggingView();
+        $query['cid'] = Uuid::fromString($query['cid'])->toString();
+        $query['iid'] = Uuid::fromString(NetworkImpression::first()->impression_id)->toString();
+
+        $response = $this->get(self::buildLogViewUri($banner->uuid, $query));
+
+        $response->assertStatus(Response::HTTP_FOUND);
+        $response->assertHeader('Location');
+        $location = $response->headers->get('Location');
+        self::assertStringStartsWith('https://example.com/view', $location);
+        parse_str(parse_url($location, PHP_URL_QUERY), $locationQuery);
+        foreach (['cid', 'ctx', 'iid', 'pto', 'pid'] as $key) {
+            self::assertArrayHasKey($key, $locationQuery);
+        }
+        self::assertEquals('13245679-8013-2456-7980-132456798012', $locationQuery['cid']);
+        self::assertEquals('0001-00000005-CBCA', $locationQuery['pto']);
+    }
+
+    public function testLogNetworkViewWithoutContextFailWhileInvalidZoneId(): void
+    {
+        [$query, $banner, $zone] = self::initBeforeLoggingView();
+        unset($query['ctx']);
+        $query['zid'] = 'invalid';
+
+        $response = $this->get(self::buildLogViewUri($banner->uuid, $query));
+
+        $response->assertStatus(Response::HTTP_BAD_REQUEST);
+    }
+
+    public function testLogNetworkViewWhileCaseIdAndImpressionIdAndZoneIdAreUuidV4(): void
+    {
+        [$query, $banner, $zone] = self::initBeforeLoggingView();
+        $query['cid'] = Uuid::fromString($query['cid'])->toString();
+        $query['iid'] = Uuid::fromString(NetworkImpression::first()->impression_id)->toString();
+        $query['zid'] = Uuid::fromString($zone->uuid)->toString();
+
+        $response = $this->get(self::buildLogViewUri($banner->uuid, $query));
+
+        $response->assertStatus(Response::HTTP_FOUND);
+        $response->assertHeader('Location');
+        $location = $response->headers->get('Location');
+        self::assertStringStartsWith('https://example.com/view', $location);
+        parse_str(parse_url($location, PHP_URL_QUERY), $locationQuery);
+        foreach (['cid', 'ctx', 'iid', 'pto', 'pid'] as $key) {
+            self::assertArrayHasKey($key, $locationQuery);
+        }
+        self::assertEquals('13245679-8013-2456-7980-132456798012', $locationQuery['cid']);
+        self::assertEquals('0001-00000005-CBCA', $locationQuery['pto']);
+    }
+
+    public function testLogNetworkViewFailWhileImpressionIdIsMissing(): void
+    {
+        [$query, $banner, $zone] = self::initBeforeLoggingView();
+        unset($query['iid']);
+
+        $response = $this->get(self::buildLogViewUri($banner->uuid, $query));
+
+        $response->assertStatus(Response::HTTP_BAD_REQUEST);
+    }
+
+    public function testLogNetworkViewFailWhileImpressionIdIsInvalid(): void
+    {
+        [$query, $banner, $zone] = self::initBeforeLoggingView();
+        $query['iid'] = '0123456789ABCDEF0123456789ABCDEF';
+
+        $response = $this->get(self::buildLogViewUri($banner->uuid, $query));
+
+        $response->assertStatus(Response::HTTP_NOT_FOUND);
+    }
+
     private static function findJsonData(array $merge = []): array
     {
         return array_merge(
@@ -687,5 +860,82 @@ final class SupplyControllerTest extends TestCase
             'width' => '300',
             'height' => '250',
         ], $merge);
+    }
+
+    private static function buildLogClickUri(string $bannerId, ?array $query = null): string
+    {
+        $uri = sprintf('/l/n/click/%s', $bannerId);
+        if (null !== $query) {
+            $uri .= '?' . http_build_query($query);
+        }
+        return $uri;
+    }
+
+    private static function buildLogViewUri(string $bannerId, ?array $query = null): string
+    {
+        $uri = sprintf('/l/n/view/%s', $bannerId);
+        if (null !== $query) {
+            $uri .= '?' . http_build_query($query);
+        }
+        return $uri;
+    }
+
+    private static function initBeforeLoggingClick(): array
+    {
+        $arr = self::initBeforeLoggingView();
+        $query = $arr[0];
+        $banner = $arr[1];
+        $zone = $arr[2];
+        NetworkCase::factory()->create([
+            'banner_id' => $banner->uuid,
+            'case_id' => $query['cid'],
+            'network_impression_id' => NetworkImpression::firstOrFail()->id,
+            'publisher_id' => $zone->site->user->uuid,
+            'site_id' => $zone->site->uuid,
+            'zone_id' => $zone->uuid,
+        ]);
+        return $arr;
+    }
+
+    private static function initBeforeLoggingView(): array
+    {
+        /** @var NetworkImpression $impression */
+        $impression = NetworkImpression::factory()->create();
+        /** @var Site $site */
+        $site = Site::factory()->create(['user_id' => User::factory()->create()]);
+        /** @var Zone $zone */
+        $zone = Zone::factory()->create(['site_id' => $site]);
+        $campaign = NetworkCampaign::factory()->create();
+        /** @var NetworkBanner $banner */
+        $banner = NetworkBanner::factory()->create([
+            'network_campaign_id' => $campaign,
+            'view_url' => 'https://example.com/view',
+        ]);
+        $iid = Utils::base64UrlEncodeWithChecksumFromBinUuidString(hex2bin($impression->impression_id));
+        $ctx = Utils::encodeZones(
+            [
+                'page' => [
+                    'iid' => $iid,
+                    'frame' => 0,
+                    'width' => 1024,
+                    'height' => 768,
+                    'url' => 'https://adshares.net',
+                    'keywords' => '',
+                    'metamask' => 0,
+                    'ref' => '',
+                    'pop' => 0,
+                    'zone' => $zone->uuid,
+                    'options' => '[]',
+                ],
+            ]
+        );
+        $redirectUrl = Utils::urlSafeBase64Encode($banner->view_url);
+        $query = [
+            'cid' => '13245679801324567980132456798012',
+            'ctx' => $ctx,
+            'iid' => $iid,
+            'r' => $redirectUrl,
+        ];
+        return [$query, $banner, $zone];
     }
 }